+++ .teamcity/settings.kts
@@ -139,6 +139,8 @@ object RunAll : BuildType({
         }
         snapshot(Validations_LintTest) {
         }
+        snapshot(Validations_ScreenshotTests) {
+        }
     }
 })
 
@@ -455,7 +457,7 @@ object SeleniumTesting_Publish : BuildType({
             id = "RUNNER_3"
             toolPath = "%teamcity.tool.NuGet.CommandLine.4.9.2%"
             packages = "packages/react-ui-testing/Output/*.nupkg"
-            serverUrl = "https://api.nuget.org/v3/index.json"
+            serverUrl = "%env.NUGET_SOURCE%"
             apiKey = "credentialsJSON:9c594bdb-ecae-416b-ab54-3b85ce110c13"
         }
     }
@@ -525,7 +527,8 @@ object Validations : Project({
     buildType(Validations_Build)
     buildType(Validations_LintTest)
     buildType(Validations_Publish)
-    buildTypesOrder = arrayListOf(Validations_LintTest, Validations_Build, Validations_Publish)
+    buildType(Validations_ScreenshotTests)
+    buildTypesOrder = arrayListOf(Validations_LintTest, Validations_Build, Validations_Publish, Validations_ScreenshotTests)
 })
 
 object Validations_Build : BuildType({
@@ -698,9 +701,40 @@ object Validations_Publish : BuildType({
 object Validations_ScreenshotTests : BuildType({
     name = "Screenshot tests"
 
+    artifactRules = ""
+
+    maxRunningBuilds = 2
+
     vcs {
         root(DslContext.settingsRoot)
     }
 
-    disableSettings("VCS_TRIGGER")
+    steps {
+        step {
+          name = "Install"
+          id = "RUNNER_1"
+          type = "jonnyzzz.yarn"
+          param("yarn_commands", "install")
+        }
+        step {
+          name = "Build Storybook"
+          id = "RUNNER_2"
+          type = "jonnyzzz.yarn"
+          param("yarn_commands", "workspace react-ui-validations storybook:build")
+        }
+        script {
+            name = "Start"
+            id = "RUNNER_3"
+            scriptContent = """
+                start /b yarn workspace react-ui-validations storybook:serve
+                ping 127.0.0.1 -n 11
+            """.trimIndent()
+        }
+        step {
+          name = "Test UI"
+          id = "RUNNER_4"
+          type = "jonnyzzz.yarn"
+          param("yarn_commands", "workspace react-ui-validations creevey")
+        }
+    }
 })