+++ packages/react-ui-validations/selenium-tests/ValidationTests/Storybook/Sync/PreinvalidSubmitValidation.cs
@@ -10,7 +10,7 @@ namespace SKBKontur.ValidationTests.Storybook.Sync
         [Test]
         public void TestValidByDefault()
         {
-            var page = new SingleInputPage(GetWebDriver()).WaitReady();
+            var page = new SingleInputPage(WebDriver).WaitReady();
             page.Input.Value.Wait().EqualTo("bad");
             page.Input.WaitNoError();
             page.ValidationState.WaitText("none");
@@ -20,7 +20,7 @@ namespace SKBKontur.ValidationTests.Storybook.Sync
         [Test]
         public void TestValidateOnSubmit()
         {
-            var page = new SingleInputPage(GetWebDriver()).WaitReady();
+            var page = new SingleInputPage(WebDriver).WaitReady();
             page.Input.Value.Wait().EqualTo("bad");
 
             page.SubmitButton.Click();